backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "business-chatbot-backend"
version = "0.1.0"
description = "LangGraph-powered backend for the AI webchat agent."
authors = [
    { name = "Business Chatbot Team" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.30.0",
    "langchain-core>=0.2.30",
    "langchain-community>=0.2.30",
    "langchain-openai>=0.1.11",
    "langgraph>=0.1.21",
    "qdrant-client>=1.9.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "loguru>=0.7.2",
    "tenacity>=8.2.3",
    "pyjwt>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
]

[tool.setuptools.packages.find]
where = ["app"]
